refactor: Break up main.go into logical packages for better maintainability
  - Extract types into internal/types package (NodegroupInfo, AMIStatus, etc.)
  - Create internal/aws package for AWS operations (cluster, nodegroup, ami)
  - Move CLI commands to internal/commands package (list, update, version)
  - Add internal/monitoring package for update progress tracking
  - Create internal/ui package for output formatting and status display
  - Add internal/dryrun package for dry run functionality
  - Reduce main.go from 1256 lines to 27 lines (98% reduction)
  - Maintain all existing functionality and test coverage
  - Follow Go project organization best practices with internal packages
